Update Nomor 3.py

+++ Nomor 3.py
@@ -1,9 +1,9 @@
 # bab1_Projek_python_nomor_3
 # Meminta input panjang dari pengguna
-panjang = float(input("Masukkan panjang persegi panjang: "))
+panjang = int(input("Masukkan panjang persegi panjang: "))
 
 # Meminta input lebar dari pengguna
-lebar = float(input("Masukkan lebar persegi panjang: "))
+lebar = int(input("Masukkan lebar persegi panjang: "))
 
 # Menghitung luas persegi panjang
 luas = panjang * lebar
